Join feeds2html packages
